requirements.txt
certifi==2022.6.15.1
charset-normalizer==2.1.1
flatten-json==0.1.13
idna==3.3
magicseaweed==1.0.3
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
